requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
openai==1.3.7
google-generativeai==0.3.2
httpx==0.25.2
pyyaml==6.0.1
aiofiles==23.2.0
aiosqlite==0.19.0
apscheduler==3.10.4
aiohttp==3.9.1